Modified toolchanger
Removed the bottom fixing screw for the cover and replaced it with 2 on the sides to increase clearance for the z probe if the bed is really tilted.
Increased the inner diameter by 1mm as the 30t gear is a snug fit.
Added some flats where the motor sits to make sure it fits.